fix(Catalog): Reset Catalogs upon resource switch
